minor performance improvements

+++ visualvm/libs.profiler/lib.profiler.ui/src/org/graalvm/visualvm/lib/ui/charts/PieChart.java
@@ -172,7 +172,7 @@ public class PieChart extends JComponent implements ComponentListener, ChartMode
         int[] items = new int[selectedItems.size()];
 
         for (int i = 0; i < selectedItems.size(); i++) {
-            items[i] = ((Integer) selectedItems.get(i)).intValue();
+            items[i] = selectedItems.get(i).intValue();
         }
 
         return items;
